refactor: move birdeye fetcher to utils
